requirements.txt
playwright>=1.41.0
playwright-stealth>=1.0.6
python-dotenv>=1.0.1
schedule>=1.2.1
